% Auto-generated: do not edit by hand
\name{chartLegend}

\alias{chartLegend}

\title{ChartLegend component}

\description{

}

\usage{
chartLegend(key=NULL, ref=NULL)
}

\arguments{
\item{key}{Character | numeric. }

\item{ref}{Character. Allows getting a ref to the component instance.
Once the component unmounts, React will set `ref.current` to `null`
(or call the ref with `null` if you passed a callback ref).
@,see,,{@link ,https://react.dev/learn/referencing-values-with-refs#refs-and-the-dom React Docs,}}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

